fix:停机处理接口

+++ screening/src/main/java/com/ht/screening/dto/AbnormalShutdownDto.java
@@ -34,7 +34,7 @@ public class AbnormalShutdownDto implements Serializable {
     /**
      * 断纤长度
      */
-    Long dqcd;
+    Double dqcd;
 
     /**
      * 用户编号
